+++ tests/test_configuration.py
@@ -482,6 +482,7 @@ class TestDefaultConfigurationTemplate:
                 "rest-api-timeout-slow-endpoints": 60.0,
                 "mpd-timeout": 5.0,
                 "rest-api-sleep-before-next-call": 2.0,
+                "rest-api-retries-on-unexpected-state": 3,
             },
             "miscellaneous": {
                 "add-cover-and-metadata": True,